+++ src/hooks/useEdgeRule.ts
@@ -26,8 +26,7 @@ export function useEdgeRule<ND = unknown, ED = unknown>() {
       },
     },
     beforeUnmount() {
-      // TODO: once 4.6.2 is shipped, uncomment this
-      // styleRule.destroy();
+      styleRule.destroy();
     },
     mounted() {
       styleRule = this.ogma.styles.addEdgeRule(
@@ -42,5 +41,8 @@ export function useEdgeRule<ND = unknown, ED = unknown>() {
       }
       );
     },
+    render() {
+      return null;
+    }
   });
 }
